+++ src/main/java/de/slevermann/streaminggate/StreamingGateApplication.java
@@ -1,9 +1,22 @@
 package de.slevermann.streaminggate;
 
+import de.slevermann.streaminggate.config.DbConfig;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfigurationExcludeFilter;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.boot.context.TypeExcludeFilter;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
+import org.springframework.context.annotation.Import;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-@SpringBootApplication
+@Import(DbConfig.class)
+@ComponentScan(excludeFilters = {@ComponentScan.Filter(type = FilterType.CUSTOM, classes = TypeExcludeFilter.class),
+        @ComponentScan.Filter(type = FilterType.CUSTOM, classes = AutoConfigurationExcludeFilter.class)})
+@SpringBootConfiguration
+@EnableJpaRepositories
+@EntityScan
 public class StreamingGateApplication {
     public static void main(String[] args) {
         SpringApplication.run(StreamingGateApplication.class, args);