+++ src/main/java/de/slevermann/streaminggate/services/AuthService.java
@@ -1,7 +1,9 @@
 package de.slevermann.streaminggate.services;
 
 import org.springframework.lang.NonNull;
+import org.springframework.stereotype.Service;
 
+@Service
 public interface AuthService {
 
     @NonNull