+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## [Unreleased]
-### Removed
-- Unused User class
+### Added
+- Possibility to store users in a database, rather than a flat file
+### Changed
+- The config keys for users are now ```streaminggate.users.passKeys``` rather than ```streaminggate.users```
 
 ## [1.0.0] 2020-03-06
 ### Added